+++ masspcf/norms.py
@@ -13,8 +13,8 @@
 #    limitations under the License.
 
 from . import _mpcf_cpp as cpp
-from .tensor import PcfContainerLike, TensorType, _to_tensor_pcf, _get_backend
-from .typing import float32, float64
+from .tensor import PcfContainerLike, _to_tensor_pcf, _get_backend
+from .typing import pcf32, pcf64
 from .np_support import numpy_type
 from .async_task import _wait_for_task
 
@@ -22,8 +22,8 @@ import numpy as np
 
 
 def _get_norms_backend(fs) -> cpp.Norms_f32_f32 | cpp.Norms_f64_f64:
-    mapping = { (TensorType.PCF, float32): cpp.Norms_f32_f32,
-                (TensorType.PCF, float64): cpp.Norms_f64_f64 }
+    mapping = { pcf32: cpp.Norms_f32_f32,
+                pcf64: cpp.Norms_f64_f64 }
 
     return _get_backend(fs, mapping)