+++ test/python/test_numpy.py
@@ -8,23 +8,23 @@ def test_np_to_pcf_has_correct_type():
 
     f32 = mpcf.Pcf(X32)
     assert isinstance(f32._data, cpp.Pcf_f32_f32)
-    assert f32.ttype == mpcf.float32
-    assert f32.vtype == mpcf.float32
+    assert f32.ttype == mpcf.f32
+    assert f32.vtype == mpcf.f32
 
     f64 = mpcf.Pcf(X64)
     assert isinstance(f64._data, cpp.Pcf_f64_f64)
-    assert f64.ttype == mpcf.float64
-    assert f64.vtype == mpcf.float64
+    assert f64.ttype == mpcf.f64
+    assert f64.vtype == mpcf.f64
 
-    f32_64 = mpcf.Pcf(X32, dtype=mpcf.float64)
+    f32_64 = mpcf.Pcf(X32, dtype=mpcf.pcf64)
     assert isinstance(f32_64._data, cpp.Pcf_f64_f64)
-    assert f32_64.ttype == mpcf.float64
-    assert f32_64.vtype == mpcf.float64
+    assert f32_64.ttype == mpcf.f64
+    assert f32_64.vtype == mpcf.f64
 
-    f64_32 = mpcf.Pcf(X64, dtype=mpcf.float32)
+    f64_32 = mpcf.Pcf(X64, dtype=mpcf.pcf32)
     assert isinstance(f64_32._data, cpp.Pcf_f32_f32)
-    assert f64_32.ttype == mpcf.float32
-    assert f64_32.vtype == mpcf.float32
+    assert f64_32.ttype == mpcf.f32
+    assert f64_32.vtype == mpcf.f32