+++ test/python/test_pdist.py
@@ -30,14 +30,14 @@ def test_pdist_of_one_gives_zero_1x1():
     assert D[0, 0] == 0.
 
 def test_pdist_of_two_gives_correct_output():
-    X = mpcf.zeros((2,), dtype=mpcf.float64)
+    X = mpcf.zeros((2,), dtype=mpcf.pcf64)
 
     X[0] = mpcf.Pcf(np.array([[0., 10.], [2., 5.], [3., 0.]]))
     X[1] = mpcf.Pcf(np.array([[0., 5.], [6., 0.]]))
 
-    print(f'Shp {X.shape[0]}')
-
     D = mpcf.pdist(X)
+
+    assert D.dtype == np.float64
 
     assert len(D.shape) == 2
     assert D.shape[0] == 2