+++ test/python/test_tensor.py
@@ -5,7 +5,7 @@ import masspcf._mpcf_cpp as mcpp
 import numpy as np
 
 def test_shape_dunder():
-    shp = mpcf.TShape((2, 3, 5))
+    shp = mpcf.Shape((2, 3, 5))
 
     str_ = str(shp)
     repr_ = repr(shp)
@@ -22,10 +22,10 @@ def test_shape_dunder():
         print(shp[3])
     
 def test_basic_shape_strides():
-    X = mpcf.zerosT((7, 3, 5, 6))
+    X = mpcf.zeros((7, 3, 5, 6), dtype=mpcf.f64)
 
-    assert X.shape == mpcf.TShape((7, 3, 5, 6))
-    assert X.shape != mpcf.TShape((1, 2, 3))
+    assert X.shape == mpcf.Shape((7, 3, 5, 6))
+    assert X.shape != mpcf.Shape((1, 2, 3))
 
     assert X.strides[0] == 3 * 5 * 6 * 1    # 90
     assert X.strides[1] == 5 * 6 * 1        # 30
@@ -33,19 +33,19 @@ def test_basic_shape_strides():
     assert X.strides[3] == 1                #  1
 
 def test_construct_1d_tensor():
-    s = mpcf.TShape((10))
-    X = mpcf.zerosT(s)
+    s = mpcf.Shape((10))
+    X = mpcf.zeros(s)
     
     assert X.shape == s
 
 def test_dtype_results_in_correct_type():
-    X32 = mpcf.zeros((3, 2), dtype=mpcf.float32)
+    X32 = mpcf.zeros((3, 2), dtype=mpcf.pcf32)
     assert isinstance(X32, mpcf.Pcf32Tensor)
 
     assert isinstance(X32[0, 0], mpcf.Pcf)
     assert isinstance(X32[0, 0]._data, mcpp.Pcf_f32_f32)
 
-    X64 = mpcf.zeros((3, 2), dtype=mpcf.float64)
+    X64 = mpcf.zeros((3, 2), dtype=mpcf.pcf64)
     assert isinstance(X64, mpcf.Pcf64Tensor)
 
     assert isinstance(X64[0, 0], mpcf.Pcf)